improve tests for parallelized run as they can take multiple seconds each
