+++ test/global-throttle.throttling.test.js
