+++ test/utils/assert-timespan.js
@@ -8,9 +8,8 @@
  * @param {number} [tolerance=15]
  */
 function assertTimespan (t, start, end, expected, tolerance = 15) {
-  const diff = end - start
-  const delta = Math.abs(expected - diff)
-  t.ok(delta <= (expected / 100 * tolerance), 'tolerance of ' + tolerance + '%, got ' + delta + 'ms')
+  const delta = end - start
+  t.ok(delta <= (expected + Math.floor(expected / 100 * tolerance)), `tolerance of ${tolerance}% of ${expected} ms, expected ${expected}ms ± ${Math.floor(expected / 100 * tolerance)}ms got ${delta} ms`)
 }
 
 module.exports = {
